엑셀파일 .zip 파일 변환

+++ src/main/java/com/tpay/domains/franchisee/application/SignUpService.java
@@ -58,7 +58,7 @@ public class SignUpService {
         }
 
         String encodedPassword = passwordEncoder.encode(password);
-        Double defaultBalancePercentage = (double)0;
+        double defaultBalancePercentage = 0;
         FranchiseeEntity franchiseeEntity =
             FranchiseeEntity.builder()
                 .businessNumber(request.getBusinessNumber())
